backend/requirements.txt
Django==1.7.4
MySQL-python==1.2.5
djangorestframework==3.0.4
mysql==0.0.1
wsgiref==0.1.2
